Update test4.html
It seems that walking in circles only works when first holding down the turning keys, and then holding down the forward or backward key. This may be because there is only one even handler for the key, and so pressing the turning key first and then holding forward/backward while continuing to hold the turning key will allow the turning situation to "stick" while the character moves forward. 

There are three potential methods to fix this

1) Make a second key press handler and map turning to one and directional movement to the other (so that both can be called simultaneously)

2) Make an interrupt statement within both turning and directional so that when the other is pressed, it interrupts its current process to start the other thing and goes back to the initial process (eg. turn-forward-turn or forward-turn-forward)

3) A list that will store the key pressed of the user at the end of each if statement for key control.