Update test4.html
In the last edit I moved the display value function between camera controls and render the scene. This may cause a lag, but I want to be able to display the value of the mouse position. I may move this function further down the program, even to the last item.